requirements.txt
asgiref==3.8.1
asn1crypto==1.5.1
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
contourpy==1.3.2
crispy-bootstrap4==2024.10
cryptography==45.0.3
cssselect2==0.8.0
curl_cffi==0.11.2
cvxpy==1.3.1
cycler==0.12.1
Cython==3.1.1
dj-database-url==3.0.0
Django==5.2.2
django-bootstrap4==25.1
django-crispy-forms==2.4
django-wkhtmltopdf==3.4.0
ecos==2.0.14
environs==14.2.0
fonttools==4.58.2
frozendict==2.4.6
html5lib==1.1
hurst==0.0.5
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lxml==5.4.0
MarkupSafe==3.0.2
marshmallow==4.0.0
matplotlib==3.7.1
multitasking==0.0.11
numpy==1.26.4
oscrypto==1.3.0
osqp==1.0.4
packaging==25.0
pandas==2.3.0
pandas-datareader==0.10.0
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
protobuf==6.31.1
pycparser==2.22
pyparsing==3.2.3
pypdf==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.1
requests==2.32.3
scipy==1.10.1
scs==3.2.7.post2
seaborn==0.13.2
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
webencodings==0.5.1
websockets==15.0.1
whitenoise==6.9.0
yfinance==0.2.61
